+++ morpheus/_lib/tests/io/test_data_loader.cpp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-#include "../test_morpheus.hpp"  // IWYU pragma: associated
+#include "../test_utils/common.hpp"  // IWYU pragma: associated
 #include "test_io.hpp"
 
 #include "morpheus/io/data_loader.hpp"
