+++ morpheus/_lib/tests/messages/test_dev_doc_ex3.cpp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-#include "../test_morpheus.hpp"  // IWYU pragma: associated
+#include "../test_utils/common.hpp"  // IWYU pragma: associated
 
 #include "morpheus/messages/meta.hpp"        // for MessageMeta
 #include "morpheus/messages/multi.hpp"       // for MultiMessage
