+++ morpheus/_lib/tests/modules/test_data_loader_module.cpp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-#include "../test_morpheus.hpp"  // IWYU pragma: associated
+#include "../test_utils/common.hpp"  // IWYU pragma: associated
 #include "mrc/runnable/types.hpp"
 #include "test_modules.hpp"
