+++ morpheus/_lib/tests/test_file_in_out.cpp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-#include "./test_morpheus.hpp"  // IWYU pragma: associated
+#include "./test_utils/common.hpp"  // IWYU pragma: associated
 
 #include "morpheus/io/deserializers.hpp"
 #include "morpheus/io/serializers.hpp"
