--- morpheus/_lib/tests/test_main.cpp
@@ -1,29 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2022-2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include <gflags/gflags.h>  // for ParseCommandLineFlags
-#include <glog/logging.h>
-#include <gtest/gtest.h>  // IWYU pragma: keep
-
-int main(int argc, char** argv)
-{
-    FLAGS_alsologtostderr = true;  // Log to console
-    ::google::InitGoogleLogging("morpheus::test_libmorpheus");
-    ::testing::InitGoogleTest(&argc, argv);
-    ::google::ParseCommandLineFlags(&argc, &argv, true);
-    return RUN_ALL_TESTS();
-}
